Generalization of subclass resolution
