+++ mass_api_client/resources/__init__.py
@@ -3,5 +3,5 @@ from .analysis_system import AnalysisSystem
 from .analysis_system_instance import AnalysisSystemInstance
 from .report import Report
 from .sample import Sample, DomainSample, IPSample, URISample, FileSample, ExecutableBinarySample
-from .sample_relation import DroppedBySampleRelation, ResolvedBySampleRelation, RetrievedBySampleRelation, ContactedBySampleRelation, SsdeepRelation
+from .sample_relation import SampleRelation, DroppedBySampleRelation, ResolvedBySampleRelation, RetrievedBySampleRelation, ContactedBySampleRelation, SsdeepRelation
 from .scheduled_analysis import ScheduledAnalysis
