+++ mass_api_client/resources/sample_relation.py
@@ -1,11 +1,12 @@
 from mass_api_client.schemas import DroppedBySampleRelationSchema, ResolvedBySampleRelationSchema, \
     RetrievedBySampleRelationSchema, ContactedBySampleRelationSchema, SsdeepSampleRelationSchema
-from .base import BaseResource
+from .base_with_subclasses import BaseWithSubclasses
 from .sample import Sample
 
 
-class SampleRelation(BaseResource):
+class SampleRelation(BaseWithSubclasses):
     endpoint = 'sample_relation'
+    _class_identifier = 'SampleRelation'
 
     @classmethod
     def create(cls, sample, other, **kwargs):
@@ -23,24 +24,29 @@ class SampleRelation(BaseResource):
 
 class DroppedBySampleRelation(SampleRelation):
     schema = DroppedBySampleRelationSchema()
+    _class_identifier = 'SampleRelation.DroppedBySampleRelation'
     creation_point = 'sample_relation/submit_dropped_by'
 
 
 class ResolvedBySampleRelation(SampleRelation):
     schema = ResolvedBySampleRelationSchema()
+    _class_identifier = 'SampleRelation.ResolvedBySampleRelation'
     creation_point = 'sample_relation/submit_resolved_by'
 
 
 class ContactedBySampleRelation(SampleRelation):
     schema = ContactedBySampleRelationSchema()
+    _class_identifier = 'SampleRelation.ContactedBySampleRelation'
     creation_point = 'sample_relation/submit_contacted_by'
 
 
 class RetrievedBySampleRelation(SampleRelation):
     schema = RetrievedBySampleRelationSchema()
+    _class_identifier = 'SampleRelation.RetrievedBySampleRelation'
     creation_point = 'sample_relation/submit_retrieved_by'
 
 
 class SsdeepRelation(SampleRelation):
     schema = SsdeepSampleRelationSchema()
+    _class_identifier = 'SampleRelation.SsdeepRelation'
     creation_point = 'sample_relation/submit_ssdeep'
